+++ AutoServiceManagment/AutoServiceManagment.DomainModels/Entities/CashBox.cs
@@ -19,7 +19,7 @@ namespace AutoServiceManagment.DomainModels.Entities
         public Service Service { get; set; }
 
         [Required(ErrorMessage = "Payment is required!")]
-        public decimal Payment { get; set; }
+        public double Payment { get; set; }
         [DataType(DataType.Date)]
         [DisplayFormat(DataFormatString = "{mm/dd/yyyy}")]
         public DateTime Date { get; set; } = DateTime.Now;
