+++ AutoServiceManagment/AutoServiceManagment.DomainModels/Entities/CustomerProduct.cs
@@ -10,6 +10,6 @@ namespace AutoServiceManagment.DomainModels.Entities
         public Product Product { get; set; }
         public int CustomerID { get; set; }
         public Customer Customer { get; set; }
-        public double Count { get; set; }
+        public int Count { get; set; }
     }
 }
