+++ AutoServiceManagment/AutoServiceManagment.DomainModels/Entities/Finance.cs
@@ -7,8 +7,8 @@ namespace AutoServiceManagment.DomainModels.Entities
     public class Finance : TimestampableObject, IEntity
     {
         public int Id { get; set; }
-        public decimal CommunalCost { get; set; }
-        public decimal AdditionalCost { get; set; }
+        public double CommunalCost { get; set; }
+        public double AdditionalCost { get; set; }
         [Required(ErrorMessage = "Datetime is required!")]
         public DateTime Date { get; set; }
         public bool IsDeleted { get; set; }=false;
