+++ AutoServiceManagment/AutoServiceManagment.DomainModels/Entities/Salary.cs
@@ -16,8 +16,8 @@ namespace AutoServiceManagment.DomainModels.Entities
         [DataType(DataType.Date)]
         [DisplayFormat(DataFormatString = "{mm/dd/yyyy}")]
         public DateTime Date { get; set; }= DateTime.Now;
-        public decimal Bonus { get; set; }
-        public decimal NetSalary { get; set; }
+        public double Bonus { get; set; }
+        public double NetSalary { get; set; }
         [ForeignKey(nameof(Tax))]
         public int TaxID { get; set; }
         public Tax Tax { get; set; }
